+++ app/build.gradle.kts
@@ -49,7 +49,6 @@ dependencies {
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.7.2")
     testImplementation("org.junit.jupiter:junit-jupiter-engine:5.7.2")
 
-
-    testImplementation ("io.mockk:mockk-android:1.13.5")
-    androidTestImplementation ("io.mockk:mockk-android:1.13.5")
+    testImplementation("io.mockk:mockk-android:1.13.5")
+    androidTestImplementation("io.mockk:mockk-android:1.13.5")
 }
