+++ app/src/main/java/woowacourse/movie/fragment/movielist/HomeFragment.kt
@@ -10,10 +10,10 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.activity.moviedetail.MovieDetailActivity
+import woowacourse.movie.data.MovieMockData
 import woowacourse.movie.fragment.movielist.adapter.MovieRecyclerViewAdapter
 import woowacourse.movie.model.AdUIModel
-import woowacourse.movie.data.MovieMockData
-import woowacourse.movie.movie.toPresentation
+import woowacourse.movie.model.toPresentation
 
 class HomeFragment : Fragment(), HomeContract.View {
     override lateinit var presenter: HomeContract.Presenter
